mediacapture-streams/obtaining-local-multimedia-content/navigatorusermedia/getusermedia-impossible-constraint.html

<!doctype html>
<html>
<head>
<title>Trivial mandatory constraint in getUserMedia</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#navigatorusermedia">
<link rel='stylesheet' href='/resources/testharness.css' media='all'/>
</head>
<body>
<p>When prompted, accept to share your video stream.</p>
<h1>Description</h1>
<p>This test checks that setting a trivial mandatory constraint (width &gt;=0) in getUserMedia works</p>

<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../featuredetection.js></script>
<script src=prefix.js></script>
<script>
var t = async_test("Tests that setting a trivial mandatory constraint in getUserMedia works", {timeout:10000});
t.step(function() {
  navigator.getUserMedia({video: {mandatory: {width: {min:Infinity}}}}, t.step_func(function (stream) {
    assert_unreached("a Video stream of infinite width cannot be created");
    t.done();
  }), t.step_func(function(error) {
    console.log(error);
    assert_equals(error.name, "ContraintNotSatisfied", "An impossible constraint triggers a ContraintNotSatified error");
    assert_equals(error.constraintName, "width", "The name of the not satisfied error is given in error.constraintName");
    t.done();
  }));
});
</script>
</body>
</html>
